requirements.txt
# Restaurant RAG Project Dependencies

# Core LangChain packages
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Vector store
faiss-cpu>=1.7.4

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API framework
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Optional: For Yelp API
# yelp-api>=0.1.0

# Optional: For enhanced parsing
html5lib>=1.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
